+++ jdecode_test.go
@@ -1,6 +1,7 @@
 package junosdecode_test
 
 import (
+	"errors"
 	"fmt"
 	"math/rand"
 	"testing"
@@ -42,6 +43,35 @@ func TestDecodeBadEncoded(t *testing.T) {
 	if passwordDecoded, err := junosdecode.Decode(
 		junWordCoded + string(rune(rand.Intn(26))) + string(rune(rand.Intn(26)))); err == nil { //nolint: gosec
 		t.Errorf("extra characters in junWordCoded not detected, passwordDecoded: %s", passwordDecoded)
+	}
+}
+
+func TestDecodeEmptyEncoded(t *testing.T) {
+	t.Parallel()
+
+	// empty input
+	if passwordDecoded, err := junosdecode.Decode(""); err == nil {
+		t.Errorf("empty input not detected, passwordDecoded: %s", passwordDecoded)
+	} else if !errors.Is(err, junosdecode.ErrEmptySecret) {
+		t.Errorf("got unexpected error: %s", err)
+	}
+}
+
+func TestDecodeNotEnoughCharsEncoded(t *testing.T) {
+	t.Parallel()
+
+	// MagicPrefix input
+	if passwordDecoded, err := junosdecode.Decode(junosdecode.MagicPrefix); err == nil {
+		t.Errorf("MagicPrefix input not detected, passwordDecoded: %s", passwordDecoded)
+	} else if !errors.Is(err, junosdecode.ErrNotEnoughChars) {
+		t.Errorf("got unexpected error: %s", err)
+	}
+
+	// MagicPrefix and 2 chars input
+	if passwordDecoded, err := junosdecode.Decode(junosdecode.MagicPrefix + "!!"); err == nil {
+		t.Errorf("MagicPrefix + 2 chars input not detected, passwordDecoded: %s", passwordDecoded)
+	} else if !errors.Is(err, junosdecode.ErrNotEnoughChars) {
+		t.Errorf("got unexpected error: %s", err)
 	}
 }